Crate ariel_os

Ariel OS is an operating system for secure, memory-safe, low-power Internet of Things (IoT). Supported hardware includes various 32-bit microcontrollers.

This is the API documentation for Ariel OS. Other resources available are:

📔 Extensive documentation for Ariel OS can be found in the book.

⚙️ The git repository is available on GitHub.

✨ Examples demonstrate various features of Ariel OS.

🧪 A set of test cases further verifies the capabilities of Ariel OS.

🚧 The roadmap shows the planned features for Ariel OS.

Cargo features

Ariel OS is highly modular with a significant number of features to configure the operating system.

System functionality

alloc — Enables a global system allocator.

external-interrupts — Enables GPIO interrupt support.

time — Enables timing functionality. Currently: Additionally enables the HAL-specific time driver.

csprng — Enables a cryptographically secure random number generator in the random module and support for the getrandom crate.

sensors — Enables unified support for sensors.

Network protocols

ipv4 — Enables support for IPv4.

ipv6 — Enables support for IPv6.

tcp — Enables support for TCP.

udp — Enables support for UDP.

dns — Enables support for DNS.

mdns — Enables support for mDNS.

multicast — Enables support for multicast (for both IPv4 and/or IPv6 if enabled).

coap — Enables support for CoAP.

coap-server — Enables applications to set up CoAP server handlers. See coap::coap_run().

Serial communication

i2c — Enables I2C support.

spi — Enables SPI support.

uart — Enables UART support.

usb — Enables USB support.

usb-hid — Enables USB HID support.

System configuration

The config attribute macro allows to provide configuration for specific system functionality. The features below need to be enabled so that the provided custom configuration is taken into account.

rcc-config-override — Enables custom RCC configuration (currently only available on STM32).

override-usb-config — Enables custom USB configuration.

Multicore functionality

core-affinity — Enables support for core affinities (restricting threads to specific cores).

Timer queue size configuration

Should only be enabled when threading is enabled, too. Default is 64 entries for regular boards and 8 entries for those with little ram. When more than this number of concurrent timers are used, the system will loop (and not sleep/idle anymore).

timer-generic-queue-8 — Enables support for up to 8 concurrent timers. (default on ram-tiny boards)

timer-generic-queue-16 — Enables support for up to 16 concurrent timers.

timer-generic-queue-32 — Enables support for up to 32 concurrent timers.

timer-generic-queue-64 — Enables support for up to 64 concurrent timers. (default)

timer-generic-queue-128 — Enables support for up to 128 concurrent timers.

Development and debugging

bench — Enables benchmarking facilities.

no-boards — Allows to have no boards selected, useful to run target-independent tooling.

Modules

asynch

Provides async functionality.

benchbench

Provides on-board benchmarking facilities.

bleble

Provides control over the system BLE (Bluetooth Low Energy) stack.

buildinfo

Exposes information about the build.

cell

Shareable containers.

coapcoap

A CoAP stack preconfigured for Ariel OS.

config

Provides configuration to the system and the application.

debug

Provides debug interface facilities.

delegate

Delegate or lend an object to another task.

gpio

Provides consistent GPIO access.

hal

Provides MCU-specific items.

i2ci2c

Provides support for the I2C communication bus.

identity

Access to unique identifiers provided by the device.

log

Provides logging facilities.

netnet

Provides network access.

power

Provides power management functionality.

randomrandom

Provides a seeded random number generator depending on Ariel OS’s configuration.

reexports

This module contains all third party crates as used by Ariel OS.

rt

Provides runtime helpers.

sensorssensors

Provides a sensor abstraction layer.

spispi

Provides support for the SPI communication bus.

storagestorage

Provides key-value pair persistent storage on flash.

threadthreading

Multi-threading for Ariel OS.

timetime

Provides time-related facilities.

uartuart

Provides support for UART.

usbusb

Provides USB-related items.

Attribute Macros

config

Allows to provide configuration for the associated driver during initial system configuration.

spawner

Registers a non-async function for autostart.

task

Defines an async task and optionally registers it for autostart.

threadthreading

Runs the function decorated with this attribute macro as a separate thread.