ariel_os::coap
Function coap_client

pub async fn coap_client() -> &'static CoAPRuntimeClient<'static, ariel_os_coap::::coap_client::{constant#0}>

Available on crate feature coap only.

Returns a CoAP client requester.

This asynchronously blocks until coap_run() has been called (which happens at startup when the corresponding feature coap-server is not active), and the CoAP stack is operational.

Panics

This is currently only available from the thread that hosts the network stack, and panics otherwise. This restriction will be lifted in the future (by generalization in [embedded_nal_coap] to allow different mutexes).