ariel_os

Module config

Provides configuration to the system and the application.

Macros

ipv4_addr_from_env

Reads an IPv4 address at compile time from the given environment variable, produces an Ipv4Addr.

ipv4_addr_from_env_or

Reads an IPv4 address at compile time from the given environment variable, produces an Ipv4Addr.

ipv6_addr_from_env

Reads an IPv6 address at compile time from the given environment variable, produces an Ipv6Addr.

ipv6_addr_from_env_or

Reads an IPv6 address at compile time from the given environment variable, produces an Ipv6Addr.

str_from_env

Reads a value at compile time from the given environment variable.

str_from_env_or

Reads a value at compile time from the given environment variable, with a default.