ariel_os::config
Macro str_from_env

macro_rules! str_from_env {
    ($env_var:literal, $doc:literal $(,)?) => { ... };
}

Reads a value at compile time from the given environment variable.

Produces a compile-time error if the environment variable is not found. The $doc parameter allows to provide a documentation string for this tunable. It should complete the following sentence: “This environment variable provides the $doc”.

Produces a compile-time error when option_env! does.