ariel_os::delegate

Struct Delegate

pub struct Delegate<T> { /* private fields */ }

Delegates or lends an object to another task.

This struct can be used to lend a &mut T to another task on the same executor. The other task can then call a closure on it. After that, the &mut T is returned to the original task.

Under the hood, Delegate leverages [SameExecutorCell] to ensure the delegated object stays on the same executor.

Example

static SOME_VALUE: Delegate<u32> = Delegate::new();

// in some task
async fn foo() {
  let mut my_val = 0u32;
  // SAFETY: `lend` is only called once.
  unsafe { SOME_VALUE.lend(&mut my_val).await; }
  assert_eq!(my_val, 1);
}

// in some other task
async fn bar() {
  SOME_VALUE.with(|val| *val = 1).await;
}

Implementations

impl<T> Delegate<T>

pub const fn new() -> Delegate<T>

Creates a new Delegate.

pub async unsafe fn lend<'a, 'b>(&'a self, something: &'b mut T)
where 'b: 'a,

Lends an object.

This blocks until Delegate::with() is called on the instance.

Safety

This must only be called once per Delegate instance.

pub async fn with<U>(&self, func: impl FnOnce(&mut T) -> U) -> U

Calls a closure on a lent object.

This blocks until Delegate::lend() is called on the instance.

Panics

Panics if called multiple times on the same Delegate instance.