ariel_os

Module gpio

Provides consistent GPIO access.

Note

This API does not currently provide a way of using the same GPIO pin as an input and an output alternatively. If you have a use case for this, especially if this is not regarding bit-banging, please open an issue on our repository.

Modules

input

Input-specific types.

output

Output-specific types.

Structs

Input

A GPIO input.

IntEnabledInputexternal-interrupts

A GPIO input that supports external interrupts.

Output

A GPIO output.

Enums

DriveStrength

Drive strength of an output.

Level

Digital level of an input or output.

Pull

Pull-up/pull-down resistor configuration.

Speed

Speed setting of an output.