ariel_os::gpio

Struct IntEnabledInput

pub struct IntEnabledInput { /* private fields */ }

Available on crate feature external-interrupts only.

A GPIO input that supports external interrupts.

Can be obtained with InputBuilder::build_with_interrupt().

Implementations

impl IntEnabledInput

pub fn is_high(&self) -> bool

Returns whether the input level is high.

pub fn is_low(&self) -> bool

Returns whether the input level is low.

pub fn get_level(&self) -> Level

Returns the input level.

pub async fn wait_for_high(&mut self)

Asynchronously waits until the input level is high. Returns immediately if it is already high.

pub async fn wait_for_low(&mut self)

Asynchronously waits until the input level is low. Returns immediately if it is already low.

pub async fn wait_for_rising_edge(&mut self)

Asynchronously waits for the input level to transition from low to high.

pub async fn wait_for_falling_edge(&mut self)

Asynchronously waits for the input level to transition from high to low.

pub async fn wait_for_any_edge(&mut self)

Asynchronously waits for the input level to transition from one level to the other.