ariel_os::identity
Function device_id_bytes

pub fn device_id_bytes() -> Result<impl AsRef<[u8]>, impl Error>

Obtains a unique identifier of the device in its byte serialized form.

See module level documentation for that identifier’s properties.

Errors

This function’s errors are device dependent, and range from the function being infallibly typed (where a device ID is accessible through memory mapped access) over bus errors (where a chip in a faulty state might impede communication with an identity EEPROM) to erring unconditionally (where no device ID is present or implemented).