ariel_os::reexports

Module ble

Available on crate feature ble only.

Common BLE types to be used across different HALs.

Structs

Config

Configuration for the BLE stack.

Constants

MAX_CHANNELS

Maximum number of concurrent channels to be handled by the BLE stack (not including GATT), minimum 1.

MAX_CONNS

Maximum number of concurrent connections to be handled by the BLE stack, minimum 1.

MTU

Maximum Transmission Unit (MTU) for BLE connections. 27 should work for all BLE versions.

Functions

get_ble_host_resources

Initializes the BLE host resources.

Type Aliases

BleHostResources

Alias for the BLE host resources, setting the generic parameters to the constants for convenience.