ariel_os::reexports::embassy_executor

Trait SpawnerTraceExt

pub trait SpawnerTraceExt {
    // Required method
    fn spawn_named<S>(
        &self,
        name: &'static str,
        token: SpawnToken<S>,
    ) -> Result<(), SpawnError>;
}

Extension trait adding tracing capabilities to the Spawner

This trait provides an additional method to spawn tasks with an associated name, which can be useful for debugging and tracing purposes.

Required Methods

fn spawn_named<S>( &self, name: &'static str, token: SpawnToken<S>, ) -> Result<(), SpawnError>

Spawns a new task with a specified name.

Arguments

name - Static string name to associate with the task

token - Token representing the task to spawn

Returns

Result indicating whether the spawn was successful

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.