ariel_os::reexports::embassy_net

Crate driver

embassy-net-driver

This crate contains the driver trait necessary for adding embassy-net support for a new hardware platform.

If you want to use embassy-net with already made drivers, you should depend on the main embassy-net crate, not on this crate.

If you are writing a driver, you should depend only on this crate, not on the main embassy-net crate. This will allow your driver to continue working for newer embassy-net major versions, without needing an update, if the driver trait has not had breaking changes.

See also embassy-net-driver-channel, which provides a higer-level API to construct a driver that processes packets in its own background task and communicates with the embassy-net task via packet queues for RX and TX.

Interoperability

This crate can run on any executor.

License

This work is licensed under either of

Apache License, Version 2.0 (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT license (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Structs

Capabilities

A description of device capabilities.

ChecksumCapabilities

A description of checksum behavior for every supported protocol.

Enums

Checksum

A description of checksum behavior for a particular protocol.

HardwareAddress

Representation of an hardware address, such as an Ethernet address or an IEEE802.15.4 address.

LinkState

The link state of a network device.

Traits

Driver

Main embassy-net driver API.

RxToken

A token to receive a single network packet.

TxToken

A token to transmit a single network packet.