ariel_os::reexports::embassy_net
Module tcp

Available on crate feature tcp only.

TCP sockets.

Listening

embassy-net does not have a TcpListener. Instead, individual TcpSockets can be put into listening mode by calling TcpSocket::accept.

Incoming connections when no socket is listening are rejected. To accept many incoming connections, create many sockets and put them all into listening mode.

Modules

client

TCP client compatible with embedded-nal-async traits.

Structs

TcpReader

The reader half of a TCP socket.

TcpSocket

A TCP socket.

TcpWriter

The writer half of a TCP socket.

Enums

AcceptError

Error returned by TcpSocket::accept.

ConnectError

Error returned by TcpSocket::connect.

Error

Error returned by TcpSocket read/write functions.

State

The state of a TCP socket, according to RFC 793.