ariel_os::reexports::embassy_time

Struct Ticker

pub struct Ticker { /* private fields */ }

Asynchronous stream that yields every Duration, indefinitely.

This stream will tick at uniform intervals, even if blocking work is performed between ticks.

For instance, consider the following code fragment.

use embassy_time::{Duration, Timer};

#[embassy_executor::task]
async fn ticker_example_0() {
    loop {
        foo();
        Timer::after(Duration::from_secs(1)).await;
    }
}

This fragment will not call foo every second. Instead, it will call it every second + the time it took to previously call foo.

Example using ticker, which will consistently call foo once a second.

use embassy_time::{Duration, Ticker};

#[embassy_executor::task]
async fn ticker_example_1() {
    let mut ticker = Ticker::every(Duration::from_secs(1));
    loop {
        foo();
        ticker.next().await;
    }
}

Cancel safety

It is safe to cancel waiting for the next tick, meaning no tick is lost if the Future is dropped.

Implementations

impl Ticker

pub fn every(duration: Duration) -> Ticker

Creates a new ticker that ticks at the specified duration interval.

pub fn reset(&mut self)

Resets the ticker back to its original state. This causes the ticker to go back to zero, even if the current tick isn’t over yet.

pub fn reset_at(&mut self, deadline: Instant)

Reset the ticker at the deadline. If the deadline is in the past, the ticker will fire instantly.

pub fn reset_after(&mut self, after: Duration)

Resets the ticker, after the specified duration has passed. If the specified duration is zero, the next tick will be after the duration of the ticker.

pub fn next(&mut self) -> impl Future<Output = ()> + Send + Sync

Waits for the next tick.

Cancel safety

The produced Future is cancel safe, meaning no tick is lost if the Future is dropped.