ariel_os::reexports::embassy_usb::class::dfu::consts

Struct DfuAttributes

pub struct DfuAttributes(/* private fields */);

Attributes supported by the DFU controller.

Implementations

impl DfuAttributes

pub const WILL_DETACH: DfuAttributes

Generate WillDetach sequence on bus.

pub const MANIFESTATION_TOLERANT: DfuAttributes

Device can communicate during manifestation phase.

pub const CAN_UPLOAD: DfuAttributes

Capable of upload.

pub const CAN_DOWNLOAD: DfuAttributes

Capable of download.

impl DfuAttributes

pub const fn empty() -> DfuAttributes

Get a flags value with all bits unset.

pub const fn all() -> DfuAttributes

Get a flags value with all known bits set.

pub const fn bits(&self) -> u8

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u8) -> Option<DfuAttributes>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u8) -> DfuAttributes

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u8) -> DfuAttributes

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<DfuAttributes>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: DfuAttributes) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: DfuAttributes) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: DfuAttributes)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: DfuAttributes)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: DfuAttributes)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: DfuAttributes, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: DfuAttributes) -> DfuAttributes

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: DfuAttributes) -> DfuAttributes

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: DfuAttributes) -> DfuAttributes

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: DfuAttributes) -> DfuAttributes

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> DfuAttributes

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl DfuAttributes

pub const fn iter(&self) -> Iter<DfuAttributes>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<DfuAttributes>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.