ariel_os::reexports::embassy_usb::class::uac1

Module speaker

USB Audio Class 1.0 - Speaker device

Provides a class with a single audio streaming interface (host to device), that advertises itself as a speaker. Includes explicit sample rate feedback.

Various aspects of the audio stream can be configured, for example:

sample rate

sample resolution

audio channel count and assignment

The class provides volume and mute controls for each channel.

Structs

AudioSettings

Audio settings for the feature unit.

ControlMonitor

Control status change monitor

Feedback

Used for writing sample rate information over the feedback endpoint.

Speaker

Implementation of the USB audio class 1.0.

State

Internal state for the USB Audio Class.

Stream

Used for reading audio frames.

Enums

Volume

The volume of an audio channel.