ariel_os::reexports::usbd_hid::descriptor

Module generator_prelude

Prelude for modules which use the gen_hid_descriptor macro.

Traits

AsInputReport

Report types which serialize into input reports, ready for transmission.

Serialize

A data structure that can be serialized into any data format supported by Serde.

SerializeTuple

Returned from Serializer::serialize_tuple.

SerializedDescriptor

Report types where serialized HID report descriptors are available.

Serializer

A data format that can serialize any data structure supported by Serde.

Attribute Macros

gen_hid_descriptor

Attribute to generate a HID descriptor & serialization code