ariel_os::sensors

Trait Reading

pub trait Reading: Debug {
    // Required method
    fn sample(&self) -> (ReadingChannel, Sample);

    // Provided method
    fn samples(&self) -> impl ExactSizeIterator + FusedIterator { ... }
}

Available on crate feature sensors only.

Implemented on Samples, returned by Sensor::wait_for_reading().

Required Methods

fn sample(&self) -> (ReadingChannel, Sample)

Returns the first value returned by Reading::samples().

Provided Methods

fn samples(&self) -> impl ExactSizeIterator + FusedIterator

Returns an iterator over Samples of a sensor reading.

The order of Samples is not significant, but is fixed.

For implementors

The default implementation must be overridden on types containing multiple Samples.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.