ariel_os::spi::main
Type Alias SpiDevice

pub type SpiDevice = SpiDevice<'static, CriticalSectionRawMutex, Spi, Output>;

Available on crate feature spi only.

An SPI driver implementing [embedded_hal_async::spi::SpiDevice].

Needs to be provided with an MCU-specific SPI driver tied to a specific SPI peripheral, obtainable from the [hal::spi::main] module. It also requires a gpio::Output for the chip select (CS) signal.

See [embedded_hal::spi] to learn more about the distinction between an SpiBus and an SpiDevice.

Note

Despite the driver interface being async, it may block during operations. However, it cannot block indefinitely as a timeout is implemented, either by leveraging SPI-specific hardware capabilities or through a generic software timeout.

Aliased Type

struct SpiDevice { /* private fields */ }