ariel_os::thread

Trait Arguable

pub unsafe trait Arguable {
    // Required method
    fn into_arg(self) -> usize;
}

Available on crate feature threading only.

Trait for types that can be used as argument for threads.

Safety

This trait must only be implemented on types whose binary representation fits into a single general-purpose register on all supported architectures.

Required Methods

fn into_arg(self) -> usize

Returns the ABI representation.