ariel_os::time
Struct Timer

pub struct Timer { /* private fields */ }

Available on crate feature time only.

A future that completes at a specified Instant.

Implementations

impl Timer

pub fn at(expires_at: Instant) -> Timer

Expire at specified Instant

pub fn after(duration: Duration) -> Timer

Expire after specified Duration. This can be used as a sleep abstraction.

Example:

use embassy_time::{Duration, Timer};

#[embassy_executor::task]
async fn demo_sleep_seconds() {
    // suspend this task for one second.
    Timer::after(Duration::from_secs(1)).await;
}

pub fn after_ticks(ticks: u64) -> Timer

Expire after the specified number of ticks.

This method is a convenience wrapper for calling Timer::after(Duration::from_ticks()). For more details, refer to Timer::after() and Duration::from_ticks().

pub fn after_nanos(nanos: u64) -> Timer

Expire after the specified number of nanoseconds.

This method is a convenience wrapper for calling Timer::after(Duration::from_nanos()). For more details, refer to Timer::after() and Duration::from_nanos().

pub fn after_micros(micros: u64) -> Timer

Expire after the specified number of microseconds.

This method is a convenience wrapper for calling Timer::after(Duration::from_micros()). For more details, refer to Timer::after() and Duration::from_micros().

pub fn after_millis(millis: u64) -> Timer

Expire after the specified number of milliseconds.

This method is a convenience wrapper for calling Timer::after(Duration::from_millis()). For more details, refer to Timer::after and Duration::from_millis().

pub fn after_secs(secs: u64) -> Timer

Expire after the specified number of seconds.

This method is a convenience wrapper for calling Timer::after(Duration::from_secs()). For more details, refer to Timer::after and Duration::from_secs().