Crate ariel_os

Ariel OS is an operating system for secure, memory-safe, low-power Internet of Things (IoT). Supported hardware includes various 32-bit microcontrollers.

This is the API documentation for Ariel OS. Other resources available are:

📔 Extensive documentation for Ariel OS can be found in the book.

⚙️ The git repository is available on GitHub.

✨ Examples demonstrates various features of Ariel OS.

🧪 A set of test cases further verifies the capabilities of Ariel OS.

🚧 The roadmap shows the planned features for Ariel OS.

Cargo features

Ariel OS is highly modular with a significant number of features to configure the operating system.

System functionality

alloc — Enables a global system allocator.

external-interrupts — Enables GPIO interrupt support.

time — Enables the internal executor’s timer queue, required for timer support.

csprng — Enables a cryptographically secure random number generator in the random module.

Network protocols and configuration

tcp — Enables support for TCP.

udp — Enables support for UDP.

dns — Enables support for DNS.

mdns — Enables support for mDNS.

coap — Enables support for CoAP.

coap-server — Enables applications to set up CoAP server handlers. See coap::coap_run().

network-config-static — Selects static IP configuration.

Serial communication

i2c — Enables I2C support.

spi — Enables SPI support.

usb — Enables USB support.

usb-hid — Enables USB HID support.

System configuration

The config attribute macro allows to provide configuration for specific system functionality. The features below need to be enabled so that the provided custom configuration is taken into account.

network-config-override — Enables custom network configuration.

override-usb-config — Enables custom USB configuration.

Multicore functionality

core-affinity — Enables support for core affinities (restricting threads to specific cores).

Development and debugging

bench — Enables benchmarking facilities.

no-boards — Allows to have no boards selected, useful to run target-independent tooling.

Modules

asynch

Provides async functionality.

benchbench

Provides on-board benchmarking facilities.

buildinfo

Exposes information about the build.

cell

Shareable containers.

coapcoap

A CoAP stack preconfigured for Ariel OS.

config

Provides configuration to the system and the application.

debug

Provides debug interface facilities.

delegate

Delegate or lend an object to another task.

gpio

Provides consistent GPIO access.

hal

Provides MCU-specific items.

i2ci2c

Provides support for the I2C communication bus.

identity

Access to unique identifiers provided by the device.

netnet

Provides network access.

randomrandom

Provides a seeded random number generator depending on Ariel OS’s configuration.

reexports

This module contains all third party crates as used by Ariel OS.

spispi

Provides support for the SPI communication bus.

storagestorage

Provides key-value pair persistent storage on flash.

threadthreading

Multi-threading for Ariel OS.

timetime

Provides time-related facilities.

usbusb

Provides USB-related items.

Attribute Macros

config

Allows to provide configuration for the associated driver during initial system configuration.

spawner

Registers a non-async function for autostart.

task

Defines an async task and optionally registers it for autostart.

threadthreading

Runs the function decorated with this attribute macro as a separate thread.