ariel_os
Module asynch

Provides async functionality.

This module bundles Ariel OS async functionality. It also wraps some Embassy types.

Modules

blockerthreading

Provides a block_on() function to use futures from a thread.

thread_executorexecutor-thread

Provides an executor that should run inside a thread.

Structs

SendSpawner

Handle to spawn tasks into an executor from any thread.

Spawner

Handle to spawn tasks into an executor.

Functions

spawner

Gets a spawner for the system executor.

yield_now

Yield from the current task once, allowing other tasks to run.