ariel_os
Crate coap

Available on crate feature coap only.

A CoAP stack preconfigured for Ariel OS.

This crate mainly provides easy-to-use wrappers around the coapcore crate, with presets tailored towards Ariel OS: It utilizes embassy_net to open a network accessible CoAP socket and selects [embedded_nal_coap] for CoAP over UDP, it selects ariel_os_random as a source of randomness, and [lakers_crypto_rustcrypto] for the cryptographic algorithm implementations.

Functions

coap_client

Returns a CoAP client requester.

coap_runcoap-server

Runs a CoAP server with the given handler on the system’s CoAP transports.