ariel_os::debug

Crate log

Provides debug logging facilities.

Syntax of formatting strings

The behavior of the provided logging macros depends on which Cargo feature is enabled:

When the defmt feature is enabled, defmt is used for logging.

When the log feature is enabled, log is used for logging.

Otherwise, the logging macros are no-ops.

This means that the syntax of the formatting strings differs depending on the enabled Cargo feature; please refer to the documentation of those crates for details on the supported syntax.

Modules

defmtdefmt

Selected [defmt] items.

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

info

Logs a message at the info level.

trace

Logs a message at the trace level.

warn

Logs a message at the warn level.