ariel_os::reexports::embassy_usb
Module msos

Microsoft OS Descriptors

https://docs.microsoft.com/en-us/windows-hardware/drivers/usbcon/microsoft-os-2-0-descriptors-specification

Modules

windows_version

Microsoft Windows version codes

Structs

CcgpDeviceDescriptor

Table 18. Microsoft OS 2.0 CCGP device descriptor.

CompatibleIdFeatureDescriptor

Table 13. Microsoft OS 2.0 compatible ID descriptor.

ConfigurationSubsetHeader

Table 11. Configuration subset header.

DescriptorSetHeader

Table 10. Microsoft OS 2.0 descriptor set header.

DescriptorSetInformation

Table 5. Descriptor set information structure.

FunctionSubsetHeader

Table 12. Function subset header.

MinimumRecoveryTimeDescriptor

Table 16. Microsoft OS 2.0 minimum USB recovery time descriptor.

ModelIdDescriptor

Table 17. Microsoft OS 2.0 model ID descriptor.

MsOsDescriptorSet

A serialized Microsoft OS 2.0 Descriptor set.

MsOsDescriptorWriter

Writes a Microsoft OS 2.0 Descriptor set into a buffer.

PlatformDescriptor

Table 4. Microsoft OS 2.0 platform capability descriptor header.

RegistryPropertyFeatureDescriptor

Table 14. Microsoft OS 2.0 registry property descriptor

VendorRevisionDescriptor

Table 19. Microsoft OS 2.0 vendor revision descriptor.

Enums

DescriptorType

Table 9. Microsoft OS 2.0 descriptor wDescriptorType values.

PropertyData

Data values that can be encoded into a registry property descriptor

PropertyDataType

Table 15. wPropertyDataType values for the Microsoft OS 2.0 registry property descriptor.

Traits

DeviceLevelDescriptor

A marker trait for feature descriptors that are valid at the device level.

FunctionLevelDescriptor

A marker trait for feature descriptors that are valid at the function level.