ariel_os::reexports::usbd_hid
Module hid_class

Implements HID functionality for a usb-device device.

Structs

HIDClass

HIDClass provides an interface to declare, read & write HID reports.

HidClassSettings

Used to define specialized HID device settings Most commonly used to setup Boot Mode (6KRO) or Report Mode (NKRO) keyboards. Some OSs will also respect the HID locale setting of the keyboard to help choose the OS keyboard layout.

ReportInfo

Enums

HidCountryCode

List of official USB HID country codes See (6.2.1): https://www.usb.org/sites/default/files/hid1_11.pdf

HidProtocol

Defines fixed packet format Only used if HidSubClass::Boot(1) is set See (4.3): https://www.usb.org/sites/default/files/hid1_11.pdf

HidProtocolMode

Get/Set Protocol mapping See (7.2.5 and 7.2.6): https://www.usb.org/sites/default/files/hid1_11.pdf

HidSubClass

Used to enable Boot mode descriptors for Mouse and Keyboard devices. See (4.2): https://www.usb.org/sites/default/files/hid1_11.pdf Boot mode descriptors are fixed and must follow a strict format. See (Appendix F): https://www.usb.org/sites/default/files/hid1_11.pdf

ProtocolModeConfig

It is often necessary to override OS behavior in order to get around OS (and application) level bugs. Forcing either Boot mode (6KRO) and Report mode (NKRO) are often necessary for NKRO compatible keyboards. Mice that support boot mode are not common and generally only useful for legacy OSs.

ReportType