ariel_os
Crate storage

Available on crate feature storage only.

Provides key-value pair persistent storage on flash.

Currently the same type used for serializing must be used for deserializing. While not doing so won’t cause unsafety, it might return garbage data, or panic.

Structs

PostcardValue

A [Value] serialized using Postcard.

Storage

Object holding an instance of a key-value pair storage.

Constants

DATA_BUFFER_SIZE

Data buffer length.

MAX_KEY_LEN

Maximum key length.

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Functions

erase_all

Resets the flash in the entire flash range.

get

Gets the last stored value from the flash that is associated with the given key.

insert

Stores a key-value pair into flash memory.

lock

Gets a [MutexGuard] of the global Storage object.

removeNon-context="stm32"

Deletes an item from flash.

Derive Macros

Deserialize

Serialize